+++ plugins/woocommerce-admin/packages/components/src/pill/docs/example.js
@@ -1,0 +1,10 @@
+/**
+ * Internal dependencies
+ */
+import { Pill } from '@woocommerce/components';
+
+export default () => {
+	return (
+		<Pill>Content</Pill>
+	);
+};