+++ plugins/woocommerce-admin/packages/components/src/pill/pill.js
@@ -1,0 +1,7 @@
+export function Pill( { children } ) {
+	return (
+		<span className="woocommerce-pill">
+			{ children }
+		</span>
+	);
+}