+++ plugins/woocommerce-admin/packages/components/src/pill/stories/index.js
@@ -1,0 +1,15 @@
+/**
+ * Internal dependencies
+ */
+import Pill from '../';
+
+export default {
+	title: 'WooCommerce Admin/components/Pill',
+	component: Pill,
+}
+
+export function Default() {
+	return (
+		<Pill>Content</Pill>
+	);
+}