+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fix `<SelectControl />` component to allow clicking anywhere on options in list to select.
 - Add support for `<List />` component item tags and link types.
 - Add `<List />` and `<Link />` components to Storybook.
+- Add `<Pill />` component.
 - Add `key` prop to `<List />` component items.
 
 ## Breaking Changes